Fixes https://github.com/totem/cluster-deployer/issues/42
- Add support for top level environment configuration
- Move deployment default logic to service (from task)
